+++ doc/whats_new.rst
@@ -20,13 +20,17 @@ NEW
   Use :func:`nilearn.reporting.plot_event` to visualize events file.
   The function accepts the BIDS events file read using `pandas` 
   utilities.
+- New plotting function
+  :func:`nilearn.plotting.plot_carpet` generates a "carpet plot" (also known
+  as a "Power plot" or a "grayplot"), for visualizing global patterns in
+  4D functional data over time.
 
 Fixes
 -----
 
 - :class:`nilearn.input_data.NiftiLabelsMasker` no longer ignores its `mask_img`
-- :func:`nilearn.masking.compute_brain_mask` has been replaced by
-  `:func: nilearn.masking.compute_gray_matter_mask`. Features remained the same but
+- :func:`nilearn.masking.compute_brain_mask` has replaced
+  `nilearn.masking.compute_gray_matter_mask`. Features remained the same but
   some corrections regarding its description were made in the docstring.